Do not try to hide find_package(Kokkos)
Before this commit, we had a fake "find_package(Kokkos)" to make tutorial CMakeLists.txt minimal.

However, using builtin CMake features like `CMAKE_DISABLE_FIND_PACKAGE_<>` or `CMAKE_REQUIRE_FIND_PACKAGE_<>` was not working properly.

This commit use a simple `include(../cmake/SetUpKokkos.cmake)`, and it
is this file that can be a fine example of how to integrate Kokkos in a project.